added priority qeue

+++ include/graph.h
@@ -5,7 +5,7 @@
 
 struct graph{
     uint16_t v;
-    uint16_t e;
+    uint32_t e;
     uint16_t** adj_matrix;
 };
